Organize helpers. Remove config autogen files from github
