--- subgraph/config/mainnet.json
@@ -1,7 +1,0 @@
-{
-  "network": "mainnet",
-  "everestAddress": "0xCe9795B2a8c4dB753ec5867E4475FCCD988Aa8Bd",
-  "everestAddressOld": "0x62c5191b39d993A68Dc9348a4a15454C60cb6C9C",
-  "blockNumber": "9892944",
-  "blockNumberOld": "9839401"
-}
