--- subgraph/config/ropsten.json
@@ -1,7 +1,0 @@
-{
-  "network": "ropsten",
-  "everestAddress": "0x275DB6f75F53D8E94c34797D84B1c1896043c6a3",
-  "everestAddressOld": "0xeCe52D2bA232fde90f56AAe6aBBeaCb17ef2b546",
-  "blockNumber": "7735064",
-  "blockNumberOld": "7730906"
-}
